+++ meaningless/bible_base_downloader.py
@@ -96,7 +96,8 @@ class BaseDownloader:
 
     def __init__(self, file_writing_function, translation='NIV', show_passage_numbers=True,
                  default_directory=os.getcwd(), strip_excess_whitespace=False, enable_multiprocessing=True,
-                 use_ascii_punctuation=False, file_extension='', write_key_as_string=False, get_raw_output=False):
+                 use_ascii_punctuation=False, file_extension='', write_key_as_string=False, capitalise_small_caps=False,
+                 get_raw_output=False):
         """
         :param file_writing_function: Function definition used to specify how to write to a given file.
                                       The function should only take 2 arguments, which are the file path to write to
@@ -124,6 +125,9 @@ class BaseDownloader:
         :param write_key_as_string: If True, specifies that all keys in the downloaded file are converted to strings.
                Defaults to False.
         :type write_key_as_string: bool
+        :param capitalise_small_caps: If True, preserves capitalisation of words through specific styling.
+                                      Defaults to False.
+        :type capitalise_small_caps: bool
         :param get_raw_output: When True, returns downloaded results as a dictionary instead of writing them to a file.
                                Defaults to False.
         :type get_raw_output: bool
@@ -137,6 +141,7 @@ class BaseDownloader:
         self.file_extension = file_extension
         self.file_writing_function = file_writing_function
         self.write_key_as_string = write_key_as_string
+        self.capitalise_small_caps = capitalise_small_caps
         self.get_raw_output = get_raw_output
 
     def download_passage(self, book, chapter, passage, file_path=''):
@@ -295,7 +300,8 @@ class BaseDownloader:
 
         online_bible = WebExtractor(translation=translation, show_passage_numbers=self.show_passage_numbers,
                                     output_as_list=True, strip_excess_whitespace_from_list=self.strip_excess_whitespace,
-                                    use_ascii_punctuation=self.use_ascii_punctuation)
+                                    use_ascii_punctuation=self.use_ascii_punctuation,
+                                    capitalise_small_caps=self.capitalise_small_caps)
 
         # Set up the base document with the root-level keys
         # Upon downloading a file, the top-level keys might be ordered differently to when they were inserted.