+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.4.0
 - Added `get_raw_output` flag to Base Downloader to obtain passage output as a data structure instead of writing to a local file
-- Added `capitalise_small_caps` flag to retain CSS capitalised words
+- Added `capitalise_small_caps` flag to Base Downloader and Web Extractor to retain CSS capitalised words
 - Drop library support for Python 3.10
 
 ## 1.3.0